[#204] Add version coverage for clang and gcc

+++ src/cli.c
@@ -1224,7 +1224,7 @@ get_config_key_result(char* config_key, struct json* j, uintptr_t* r, int32_t ou
    memset(server, 0, MISC_LENGTH);
    memset(key, 0, MISC_LENGTH);
 
-   for (int i = 0; i < strlen(config_key); i++)
+   for (size_t i = 0; i < strlen(config_key); i++)
    {
       if (config_key[i] == '.')
       {
